:feet: Move errors tests to tests/test_errors.py
